recordTable: Use instance instead of static, arity: int -> size_t

+++ src/RecordTable.cpp
@@ -23,9 +23,9 @@ namespace {
 /**
  * The static access function for record maps of certain arities.
  */
-RecordMap& getForArity(int arity) {
+RecordMap& getForArity(size_t arity) {
     // the static container -- filled on demand
-    static std::unordered_map<int, RecordMap> maps;
+    static std::unordered_map<size_t, RecordMap> maps;
     auto pos = maps.find(arity);
     if (pos == maps.end()) {
         maps.emplace(arity, arity);
@@ -35,17 +35,17 @@ RecordMap& getForArity(int arity) {
 }
 }  // namespace
 
-RamDomain RecordTable::pack(RamDomain* tuple, int arity) {
+RamDomain RecordTable::pack(RamDomain* tuple, size_t arity) {
     // conduct the packing
     return getForArity(arity).pack(tuple);
 }
 
-RamDomain* RecordTable::unpack(RamDomain ref, int arity) {
+RamDomain* RecordTable::unpack(RamDomain ref, size_t arity) {
     // conduct the unpacking
     return getForArity(arity).unpack(ref);
 }
 
-void RecordTable::createRecordMap(int arity) {
+void RecordTable::createRecordMap(size_t arity) {
     getForArity(arity);
 }
